remove unneeded interface
